Monster commit! Changes:
Shifted matrix code to its own subdirectory as I can see the code is going to sprawl. The top level matrix.ss is still around till I get this all sorted out.

Add more of a matrix API, in particular comprehensions.

Add a low-level routine to treat a matrix as a continuous array. Useful for the in-vector sequence.

I've just about run out of time for this library, for now. Another day, and I better just add the stuff I need and get back to my main project.